refactor: 각 기능별로 디렉토리를 분리, Mybatis 의존성 추가
