+++ src/main/java/com/teamProject/lostArkProject/calendar/controller/CalendarRestController.java
@@ -1,15 +1,15 @@
-package com.teamProject.lostArkProject.controller;
+package com.teamProject.lostArkProject.calendar.controller;
 
-import com.teamProject.lostArkProject.domain.Calendar;
-import com.teamProject.lostArkProject.dto.CalendarDTO;
-import com.teamProject.lostArkProject.service.CalendarService;
+import com.teamProject.lostArkProject.calendar.domain.Calendar;
+import com.teamProject.lostArkProject.calendar.dto.CalendarDTO;
+import com.teamProject.lostArkProject.calendar.service.CalendarService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
 import reactor.core.publisher.Mono;
 
 import java.util.List;
+import java.util.Map;
 
 @RestController
 @RequiredArgsConstructor
@@ -30,4 +30,10 @@ public class CalendarRestController {
     public Mono<List<CalendarDTO>> getCalendarDTOList() {
        return calendarService.getCalendarWithRemainTime();
     }
+
+    //@GetMapping("/remaintimes")
+    //public Mono<Map<String, Calendar>> getRemainTimes() {
+    //    return lostArkAPIService.getCalendar()
+    //            .flatMap(calendar -> lostArkAPIService.getRemainTimes());
+    //}
 }
