+++ src/main/java/com/teamProject/lostArkProject/calendar/service/CalendarConfig.java
@@ -1,21 +1,17 @@
-package com.teamProject.lostArkProject.service;
+package com.teamProject.lostArkProject.calendar.service;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.teamProject.lostArkProject.dto.CalendarDTO;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
 import org.springframework.web.reactive.function.client.WebClient;
 
-import java.util.List;
-
 
 @Configuration
 @PropertySource("classpath:config.properties")
-public class LostArkAPIConfig {
+public class CalendarConfig {
     @Value("${lostArkAPIKey}")
     private String lostArkAPIKey;
