+++ src/main/java/com/teamProject/lostArkProject/collectible/domain/CollectibleItem.java
@@ -1,4 +1,4 @@
-package com.teamProject.lostArkProject.domain;
+package com.teamProject.lostArkProject.collectible.domain;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
