+++ src/main/java/com/teamProject/lostArkProject/collectible/service/CollectibleConfig.java
@@ -1,0 +1,26 @@
+package com.teamProject.lostArkProject.collectible.service;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.web.reactive.function.client.WebClient;
+
+
+@Configuration
+@PropertySource("classpath:config.properties")
+public class CollectibleConfig {
+    @Value("${lostArkAPIKey}")
+    private String lostArkAPIKey;
+
+    @Bean
+    public WebClient webClient() {
+        return WebClient.builder()
+                .codecs(configurer -> configurer.defaultCodecs().maxInMemorySize(10 * 1024 * 1024)) // 10MB로 설정
+                .baseUrl("https://developer-lostark.game.onstove.com")
+                .defaultHeader("Authorization", "Bearer " + lostArkAPIKey)
+                .build();
+    }
+}
