+++ src/main/java/com/teamProject/lostArkProject/main/controller/MainController.java
@@ -1,11 +1,12 @@
-package com.teamProject.lostArkProject.controller;
+package com.teamProject.lostArkProject.main.controller;
 
-import com.teamProject.lostArkProject.domain.CharacterInfo;
-import com.teamProject.lostArkProject.domain.CollectibleItem;
-import com.teamProject.lostArkProject.service.LostArkAPIService;
+import com.teamProject.lostArkProject.calendar.service.CalendarService;
+import com.teamProject.lostArkProject.collectible.domain.CharacterInfo;
+import com.teamProject.lostArkProject.collectible.domain.CollectibleItem;
+import com.teamProject.lostArkProject.collectible.service.CollectibleService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
-import org.springframework.http.HttpStatusCode;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -16,12 +17,10 @@ import java.util.ArrayList;
 import java.util.List;
 
 @Controller
+@RequiredArgsConstructor
 public class MainController {
-    private final LostArkAPIService lostArkAPIService;
-
-    public MainController(LostArkAPIService lostArkAPIService) {
-        this.lostArkAPIService = lostArkAPIService;
-    }
+    private final CalendarService calendarService;
+    private final CollectibleService collectibleService;
 
     @GetMapping("/")
     public String character(Model model) {
@@ -70,7 +69,7 @@ public class MainController {
         // 클라이언트가 입력한 캐릭터 닉네임을 http 세션에 저장
         HttpSession session = request.getSession();
         session.setAttribute("nickname", characterName);
-        Mono<List<CharacterInfo>> characterInfoMono = lostArkAPIService.getCharacterInfo(characterName);
+        Mono<List<CharacterInfo>> characterInfoMono = collectibleService.getCharacterInfo(characterName);
         return characterInfoMono.flatMap(characterInfoList -> {
             model.addAttribute("characterList", characterInfoList);
             return Mono.just("characters");
@@ -81,8 +80,8 @@ public class MainController {
     @GetMapping("/calendar")
     public Mono<String> calendar(Model model) {
         return Mono.zip(
-                lostArkAPIService.getCalendar(),
-                lostArkAPIService.getRemainTimes()
+                calendarService.getCalendar(),
+                calendarService.getRemainTimes()
         ).doOnNext(tuple -> {
             model.addAttribute("calendar", tuple.getT1());
             model.addAttribute("remainTimes", tuple.getT2());
@@ -92,8 +91,8 @@ public class MainController {
     @GetMapping("/index")
     public Mono<String> home(Model model) {
         return Mono.zip(
-                lostArkAPIService.getCalendar(),
-                lostArkAPIService.getRemainTimes()
+                calendarService.getCalendar(),
+                calendarService.getRemainTimes()
         ).doOnNext(tuple -> {
             model.addAttribute("calendar", tuple.getT1());
             model.addAttribute("remainTimes", tuple.getT2());
@@ -104,7 +103,7 @@ public class MainController {
     @GetMapping("/collectible")
     public Mono<String> getCharacterCollectable(Model model, HttpSession session) {
         String characterName = (String) session.getAttribute("nickname"); // http 세션에서 가져온 닉네임
-        Mono<List<CollectibleItem>> collectibleItemMono = lostArkAPIService.getCharacterCollectible(characterName);
+        Mono<List<CollectibleItem>> collectibleItemMono = collectibleService.getCharacterCollectible(characterName);
         return collectibleItemMono.flatMap(collectibleItemList -> {
             model.addAttribute("collectibleItemList", collectibleItemList);
             return Mono.just("project/collectible"); // 결과 뷰로 이동
