+++ src/main/resources/mapper/Mapper.xml
@@ -1,0 +1,35 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.teamProject.lostArkProject.calendar.domain.Calendar">
+    <insert id="insertBoard" parameterType="com.example.mybatisspring.dto.Board">
+        insert into board(user_id,title,content) values (#{userId},#{title},#{content})
+    </insert>
+    <select id="countBoardList" resultType="int">
+        select count(*) from board
+    </select>
+    <select id="countMyBoardList" parameterType="int" resultType="int">
+        select count(*) from board where user_id = #{userId}
+    </select>
+    <select id="getBoardList" parameterType="com.example.mybatisspring.dto.PageRequest">
+        select id,title,content,user_id as userId,created_at as time from board
+        order by id desc
+            limit #{size} offset #{offset}
+    </select>
+    <select id="getMyBoardList" parameterType="int">
+        select id,title,content,user_id as userId,created_at as time from board
+        where user_id = #{userId}
+        order by id desc
+            limit #{size} offset #{offset}
+    </select>
+    <select id="getBoard" parameterType="int">
+        select id,title,content,user_id as userId,created_at as time from board where id = #{id}
+    </select>
+    <update id="updateBoard" parameterType="com.example.mybatisspring.dto.Board">
+        update board set title = #{title} , content = #{content} where id = #{id}
+    </update>
+    <delete id="deleteBoard" parameterType="int">
+        delete from board where id = #{id}
+    </delete>
+</mapper>
